refactor(ui): move inline components to own dir
