+++ components/ui/index.ts
@@ -11,7 +11,3 @@ export * from './Overlay'
 export * from './DocsPagination'
 export * from './BlogPagination'
 export * from './DynamicLink'
-export * from './EditToggle'
-export * from './DiscardButton'
-export * from './InlineWysiwyg'
-export * from './InlineTextareaField'
